fix: preserve current database contracts after rebase
Amp-Thread-ID: https://ampcode.com/threads/T-019fc7c2-4e78-735f-862b-a72107e4eee5

+++ internal/services/analytics/executor.go
@@ -37,7 +37,7 @@ func Execute(ctx context.Context, manager ConnectionManager, req ExecuteRequest)
 		AllowedTables:     req.AllowedTables,
 		SecurityFilters:   append([]queryparser.SecurityFilter(nil), req.SecurityFilters...),
 		Limit:             int(settings.ClickhouseMaxQueryResultRows),
-		QueryRangeDaysMax: settings.QuotaLogsRetentionDays,
+		QueryRangeDaysMax: int32(settings.QuotaLogsRetentionDays),
 	})
 
 	parsedQuery, err := parser.Parse(ctx, req.Query)
